[#179] BaseView 작업
### Description

- ViewPresentable 프로토콜에서 BaseView를 하나 만들고 override해서 사용

Resolve #179
