--- SobokSobok/SobokSobok/Common/Protocol/ViewPresentable.swift
@@ -1,13 +1,0 @@
-//
-//  ViewPresentable.swift
-//  SobokSobok
-//
-//  Created by 김승찬 on 2022/04/29.
-//
-
-import Foundation
-
-protocol ViewPresentable {
-    func setupView()
-    func setupConstraints()
-}
