+++ SobokSobok/SobokSobok/Presentation/Pill/PillFirst/View/PillPeriodView.swift
@@ -10,7 +10,7 @@ import UIKit
 import SnapKit
 import Then
 
-final class PillPeriodView: UIView, ViewPresentable {
+final class PillPeriodView: BaseView {
     
     let backgroundView = UIView().then {
         $0.makeRounded(radius: 15)
@@ -25,24 +25,13 @@ final class PillPeriodView: UIView, ViewPresentable {
         $0.titleLabel?.font = UIFont.font(.pretendardSemibold, ofSize: 18)
     }
     
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupView()
-        setupConstraints()
-    }
-    
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-    }
-    
-    func setupView() {
+    override func setupView() {
         [backgroundView, pickerView, confirmButton].forEach {
             addSubview($0)
         }
     }
     
-    func setupConstraints() {
+    override func setupConstraints() {
         backgroundView.snp.makeConstraints {
             $0.height.equalTo(UIScreen.main.bounds.height * 0.3)
             $0.leading.trailing.equalToSuperview()
